refactor(activerecord): move MySQL explain clause logic to AbstractMysqlAdapter so Trilogy inherits it
